Improved printed output, InfluxDB connection testing and loading of last_run.txt

+++ dashboard-writer/utils.py
@@ -12,22 +12,25 @@ def write_influx(name, df):
         url=inputs.influx_url, token=inputs.token, org=inputs.org_id, debug=False
     )
 
-    _write_client = client.write_api(
-        write_options=WriteOptions(
-            batch_size=5000,
-            flush_interval=10_000,
-            jitter_interval=2_000,
-            retry_interval=5_000,
+    if client.health().status == "pass":
+        _write_client = client.write_api(
+            write_options=WriteOptions(
+                batch_size=5000,
+                flush_interval=10_000,
+                jitter_interval=2_000,
+                retry_interval=5_000,
+            )
         )
-    )
 
-    _write_client.write(
-        inputs.influx_bucket, record=df, data_frame_measurement_name=name,
-    )
+        _write_client.write(
+            inputs.influx_bucket, record=df, data_frame_measurement_name=name,
+        )
 
-    print(f"{len(df.index)} records written to InfluxDB")
-    _write_client.__del__()
-    client.__del__()
+        print(f"- SUCCESS: {len(df.index)} records of {name} written to InfluxDB\n\n")
+        _write_client.__del__()
+        client.__del__()
+    else:
+        print("- WARNING: Unable to connect to InfluxDB - please check credentials\n\n")
 
 
 def delete_influx(name):
@@ -40,17 +43,22 @@ def delete_influx(name):
         url=inputs.influx_url, token=inputs.token, org=inputs.org_id, debug=False
     )
 
-    start = "1970-01-01T00:00:00Z"
-    stop = "2099-01-01T00:00:00Z"
+    if client.health().status == "pass":
+        start = "1970-01-01T00:00:00Z"
+        stop = "2099-01-01T00:00:00Z"
 
-    delete_api = client.delete_api()
-    delete_api.delete(
-        start,
-        stop,
-        f'_measurement="{name}"',
-        bucket=inputs.influx_bucket,
-        org=inputs.org_id,
-    )
+        delete_api = client.delete_api()
+        delete_api.delete(
+            start,
+            stop,
+            f'_measurement="{name}"',
+            bucket=inputs.influx_bucket,
+            org=inputs.org_id,
+        )
+    else:
+        print(
+            "\nWARNING: Unable to connect to InfluxDB - please check your credentials\n\n"
+        )
 
 
 def setup_fs_s3():
@@ -87,14 +95,19 @@ def load_last_run(f_path):
     from datetime import datetime, timedelta, timezone
 
     fmt = "%Y-%m-%d %H:%M:%S"
+
     try:
         with open(f_path, mode="r") as file:
-            last_run = datetime.strptime(file.read(), fmt)
-            print(f"{f_path} found - loading data from {last_run}")
+            file_data = file.read()
+    except:
+        print(f"Warning: Unable to load file from {f_path}")
 
+    try:
+        last_run = datetime.strptime(file_data.replace("\n", ""), fmt)
+        print(f"{f_path} found - loading data from {last_run}")
     except:
         last_run = datetime.utcnow() - timedelta(days=7)
-        print(f"{f_path} not found - loading data from {last_run} (7 days ago)")
+        print(f"{f_path} is invalid - loading data from {last_run} (7 days ago)")
 
     return last_run.replace(tzinfo=timezone.utc)
